Move to the new syntax parser model
