+++ src/Raven.CodeAnalysis/Syntax/InternalSyntax/Parser/Parsers/NameSyntaxParser.cs
@@ -1,0 +1,66 @@
+namespace Raven.CodeAnalysis.Syntax.InternalSyntax.Parser;
+
+using System;
+
+using static Raven.CodeAnalysis.Syntax.InternalSyntax.SyntaxFactory;
+
+internal class NameSyntaxParser : SyntaxParser
+{
+    public NameSyntaxParser(ParseContext parent) : base(parent)
+    {
+
+    }
+
+
+    public NameSyntax ParseName()
+    {
+        NameSyntax left = ParseSimpleName();
+
+        while (ConsumeToken(SyntaxKind.DotToken, out var dotToken))
+        {
+            left = QualifiedName(left, dotToken, ParseSimpleName());
+        }
+
+        return left;
+    }
+
+    public TypeSyntax ParseTypeName()
+    {
+        var peek = PeekToken();
+        if (IsPredefinedTypeKeyword(peek))
+        {
+            ReadToken();
+
+            return PredefinedType(peek);
+        }
+
+        NameSyntax left = ParseSimpleName();
+
+        while (ConsumeToken(SyntaxKind.DotToken, out var dotToken))
+        {
+            left = QualifiedName(left, dotToken, ParseSimpleName());
+        }
+
+        return left;
+    }
+
+    public IdentifierNameSyntax ParseSimpleName()
+    {
+        var token = ReadToken();
+        return IdentifierName(token);
+    }
+
+    private bool IsPredefinedTypeKeyword(SyntaxToken token)
+    {
+        switch (token.Kind)
+        {
+            case SyntaxKind.StringKeyword:
+            case SyntaxKind.BoolKeyword:
+            case SyntaxKind.CharKeyword:
+            case SyntaxKind.IntKeyword:
+                return true;
+        }
+
+        return false;
+    }
+}
